WIP: Implement centralised error handling - Create helper functions to send and render dynamic error messages to   error.ejs (renamed from 404.ejs). - Implement the functionality into the app.get("/books/:id") in api.js and app.get("/books/edit/:id") in server.js. - Improve error.ejs to display errors and status codes dynamically as   well accounting for catch all middleware that does not send any   objects. - Functionality still pending for the other routes.
