+++ book-notes-project/backend/helper-functions/renderError.js
@@ -1,0 +1,8 @@
+function renderError(res, status, error) {
+    return res.status(status).render("error.ejs", {
+        status,
+        error
+    })
+}
+
+export default renderError;
