+++ book-notes-project/backend/helper-functions/sendError.js
@@ -1,0 +1,11 @@
+function sendError(res, status, message, details = []) {
+    res.status(status).json({
+        error: {
+            message,
+            details
+        }
+    })
+}
+
+
+export default sendError;
