+++ book-notes-project/backend/helper-functions/validator.js
