+++ book-notes-project/views/error.ejs
@@ -13,17 +13,22 @@
     <a href="/"><div>Virtual Bookroom</div></a>
 </header>
 <main id="page-404">
-    <h1>404</h1>
-    <% if (locals.errors && locals.errors.length > 0) { %>
-        <% locals.errors.forEach(err => { %>
-            <p><%= err.msg %></p>
-        <% }) %>
+    <% if (locals.status && locals.error) { %>
+        <h1><%= status %></h1>
+        <p><%= error.message %></p>
+        <% if (error.details && error.details.length > 0) { %>
+            <% error.details.forEach(err => { %>
+                <p><%= err.msg %></p>
+            <% }) %>
+        <% } %>
+
     <% } else { %>
-        <p>The route you're looking for does not exist.</p>
-    <% } %>
+        <h1>404</h1>
+        <p>The page or route you're looking for does not exist.</p>
+        <% } %>
+
 
     <a href="/" class="btn">Return to Home</a>
 </main>
-
 </body>
 </html>
